Reorganize files using redux folder name. Add useSelector to todo-list.tsx and created initial list rendering.
