--- client/components/todos-slice.ts
@@ -1,17 +1,0 @@
-import { createSlice } from '@reduxjs/toolkit';
-
-export const todosSlice = createSlice({
-  name: 'todos',
-  initialState: {
-    todos: []
-  },
-  reducers: {
-    add: (state, action) => {
-      [...state.todos, action.payload];
-    }
-  }
-})
-
-export const { add } = todosSlice.actions;
-
-export default todosSlice.reducer;
