+++ client/index.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import App from './app';
 import { Provider } from 'react-redux';
-import store from './generate-store';
+import store from './redux/generate-store';
 
 const container = document.querySelector('#root');
 const root = ReactDOM.createRoot(container);
