+++ package.json
@@ -51,6 +51,7 @@
     "markuplint": "^2.11.0",
     "nodemon": "^2.0.20",
     "react-refresh": "^0.14.0",
+    "redux-devtools-extension": "^2.13.9",
     "style-loader": "^3.3.1",
     "stylelint": "^14.14.0",
     "stylelint-config-recommended": "^9.0.0",
